Move to using unasync to generate synchronous code
I don't want to maintain the two classes by hand so let's do what the
pros seem to and use a tool to generate the sync one from the async.
