+++ src/libopensonic/_version.py
@@ -17,4 +17,4 @@ along with py-opensonic.  If not, see <http://www.gnu.org/licenses/>
 
 
 #Semantic versioning string for the library
-__version__ = '8.1.3'
+__version__ = '9.0.0'
